Todos/Findings: maybe look at cache optimizations aka 3d blocking -> test against original functions ##### todos long term: make a boundary for pressure variance and start testing with an object in a fluid investigate the separation of data in multiple points -> separation of wet and dry nodes -> separation of data-array form the nodes -> new class for the macro values rework toLinks, is mostly in order needs more testing thou make wierd channels to should be m4 for me -> next step calculate seems superior than the lookup thou, prob comes down to cpu stalls, should also do a cache hit analysis prob to shed some light on that -> still test out with more tables prob needs a SoA structure to work porperlly thou so more a bigger thing than a smaller... ##### Actual: working on image recogination stuff apparenttly there is a color mistake in the code

+++ src/types.h
@@ -15,7 +15,7 @@ using vector_t = Eigen::Vector2d;
 
 using handle_t = uint64_t; // counter to denote a node
 
-using colour_t = uint64_t; // placeholder for colour, whatever the right dataformat may be
+using colour_t = uint32_t; // placeholder for colour, whatever the right dataformat may be
 
 // globals (defined in functions)
 extern matrix_t velocity_set;
